clean security config
