+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/config/SecurityConfig.java
@@ -1,22 +1,19 @@
 package org.orange.oie.internship2025.assetmanagementsystem.config;
-
+import org.orange.oie.internship2025.assetmanagementsystem.reposetries.UserReposetries;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.authentication.ProviderManager;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import java.util.List;
 
 import static org.springframework.security.config.Customizer.withDefaults;
@@ -24,6 +21,11 @@ import static org.springframework.security.config.Customizer.withDefaults;
 @Configuration
 @EnableWebSecurity
 public class SecurityConfig {
+    private  final UserReposetries userReposetries;
+
+    public SecurityConfig(UserReposetries userReposetries) {
+        this.userReposetries = userReposetries;
+    }
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -40,32 +42,36 @@ public class SecurityConfig {
         return http.build();
     }
 
-
     @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authConfig) throws Exception {
-        return authConfig.getAuthenticationManager();
+    public AuthenticationManager authenticationManager(
+            UserDetailsService userDetailsService,
+            PasswordEncoder passwordEncoder) {
+        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider(userDetailsService);
+        authenticationProvider.setPasswordEncoder(passwordEncoder);
+
+        return new ProviderManager(authenticationProvider);
     }
-//    @Bean
-//    public PasswordEncoder passwordEncoder() {
-//        return org.springframework.security.crypto.password.NoOpPasswordEncoder.getInstance();
-//    }//we will change it After some days
 
 
     @Bean
-public PasswordEncoder passwordEncoder() {
-    return new org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder();
-}
-/*
-String rawPassword = "123456";
-String encodedPassword = passwordEncoder.encode(rawPassword);
-user.setPassword(encodedPassword);
-userRepository.save(user);
+    public UserDetailsService userDetailsService() {
+        return username -> {
+            org.orange.oie.internship2025.assetmanagementsystem.entity.User user = userReposetries.findByEmail(username)
+                    .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
+
+            return org.springframework.security.core.userdetails.User
+                    .withUsername(user.getEmail())
+                    .password(user.getPassword())//encrypted
+                    .roles(user.getRole().getRoleType())
+                    .build();
+        };
+    }
 
 
-
-
-
-     */
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder();
+    }
 
     @Bean
     public CorsConfigurationSource corsConfigurationSource() {
