+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/entity/User.java
@@ -19,7 +19,7 @@ public class User {
 
     @Column(nullable = false, unique = true, length = 200)
     private String email;
-//hello here hello againfdafdafadfda
+
     @Column(nullable = false, length = 200)
     private String password;
