+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/reposetries/UserReposetries.java
@@ -7,4 +7,5 @@ import java.util.Optional;
 
 public interface UserReposetries extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
+    Optional<User> findByPassword(String password);
 }
