+++ src/main/resources/application.properties
@@ -1,15 +1,11 @@
 # ----------------------------------------
 # Database connection
 # ----------------------------------------
-#spring.datasource.url=jdbc:mariadb://0.0.0.0:3302/ams
-#spring.datasource.username=root
-#spring.datasource.password=rootpassword
-#spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.url=jdbc:mariadb://localhost:3306/ams
 spring.datasource.username=root
 spring.datasource.password=root
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+
 
 # (If you?re using MySQL, change to:)
 # spring.datasource.url=jdbc:mysql://localhost:3306/amsdb
