+++ src/test/resources/application-test.properties
@@ -1,0 +1,15 @@
+# ----------------------------------------
+# Database connection (Test)
+# ----------------------------------------
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.show-sql=true
+
+# ----------------------------------------
+# Flyway
+# ----------------------------------------
+spring.flyway.enabled=false
